Adding CJS in addition to ESM for package
