+++ Source/JavaScript/Applications.React/index.ts
@@ -7,6 +7,7 @@ import * as identity from './identity';
 import * as queries from './queries';
 
 export * from './ApplicationModel';
+export * from './ApplicationModelContext';
 export * from './WellKnownBindings';
 
 export {
