+++ Source/JavaScript/Applications.React/tsconfig.json
@@ -3,8 +3,12 @@
     "compilerOptions": {
         "baseUrl": ".",
         "rootDir": ".",
-        "skipLibCheck": true,
-        "outDir": "dist"
+        "outDir": "dist/esm",
+        "paths": {
+            "@cratis/applications/*": [
+                "../Applications/dist/esm/*"
+            ]
+        }
     },
     "references": [
         {
