+++ Source/JavaScript/Applications.Vite/tsconfig.json
@@ -2,7 +2,7 @@
     "extends": "../../../tsconfig.json",
     "compilerOptions": {
         "rootDir": ".",
-        "outDir": "dist"
+        "outDir": "dist/esm"
     },
     "references": [
         {
