+++ Source/JavaScript/Applications/tsconfig.json
@@ -1,10 +1,14 @@
 {
     "extends": "../../../tsconfig.json",
     "compilerOptions": {
-        "rootDir": "./",
-        "outDir": "dist"
+        "baseUrl": ".",
+        "rootDir": ".",
+        "outDir": "dist/esm"
     },
-    "references": [
+    "references": [],
+    "include": [
+        "**/*.ts",
+        "**/*.tsx"
     ],
     "exclude": [
         "dist",
